refactor: RECOMMEND_ACTIVITY_KEY type 지정
