+++ apps/web/src/common/constants/index.ts
@@ -1,5 +1,7 @@
-export const RECOMMEND_ACTIVITY_KEY_ENUM = {
+export const RECOMMEND_ACTIVITY_KEY = {
   FAVORITE_PLACE_KEY_INDEX: '1',
   FAVORTIE_ACTIVITY_KEY_INDEX: '2',
   TRAVEL_THEME_KEY_INDEX: '3',
 } as const
+
+export type RECOMMEND_ACTIVITY_KEY_TYPE = (typeof RECOMMEND_ACTIVITY_KEY)[keyof typeof RECOMMEND_ACTIVITY_KEY]
