+++ apps/web/src/pages/recommend/index.page.tsx
@@ -13,7 +13,7 @@ import {
   TravelThemeQueryKey,
 } from '@/common/api/recommend'
 import DefaultLayout from '@/common/components/Layout/DefaultLayout'
-import { RECOMMEND_ACTIVITY_KEY_ENUM } from '@/common/constants'
+import { RECOMMEND_ACTIVITY_KEY, RECOMMEND_ACTIVITY_KEY_TYPE } from '@/common/constants'
 import { useRecommendPageStore } from '@/stores'
 
 import FavoriteActivity from './FavoriteActivity'
@@ -22,17 +22,17 @@ import TravelTheme from './TravelTheme'
 
 const getTabItems = () => [
   {
-    key: RECOMMEND_ACTIVITY_KEY_ENUM.FAVORITE_PLACE_KEY_INDEX,
+    key: RECOMMEND_ACTIVITY_KEY.FAVORITE_PLACE_KEY_INDEX,
     label: `좋아하는 지역`,
     children: <FavoritePlace />,
   },
   {
-    key: RECOMMEND_ACTIVITY_KEY_ENUM.TRAVEL_THEME_KEY_INDEX,
+    key: RECOMMEND_ACTIVITY_KEY.TRAVEL_THEME_KEY_INDEX,
     label: `여행테마`,
     children: <TravelTheme />,
   },
   {
-    key: RECOMMEND_ACTIVITY_KEY_ENUM.FAVORTIE_ACTIVITY_KEY_INDEX,
+    key: RECOMMEND_ACTIVITY_KEY.FAVORTIE_ACTIVITY_KEY_INDEX,
     label: `하고싶은 활동`,
     children: <FavoriteActivity />,
   },
@@ -46,9 +46,7 @@ export const Recommend = ({ favoritePlace }: any) => {
       <Suspense fallback={<div />}>
         <Tab
           activeKey={tabIndex}
-          onChange={(key) =>
-            handleChangeTabIndex(key as (typeof RECOMMEND_ACTIVITY_KEY_ENUM)[keyof typeof RECOMMEND_ACTIVITY_KEY_ENUM])
-          }
+          onChange={(key) => handleChangeTabIndex(key as RECOMMEND_ACTIVITY_KEY_TYPE)}
           className="mt-30pxr"
           defaultActiveKey="1"
           items={getTabItems()}
